Split content-items actions & reducer up into multiple files, for increased readability.
