+++ app/modules/content-items/actions/add.js
@@ -1,0 +1,19 @@
+// @flow
+
+import * as t from '../actionTypes';
+import type { ContentItemType } from '../model';
+
+const add = (
+  type: ContentItemType,
+  props: t.ActionPayloadProps,
+): t.AddAction => {
+  return {
+    type: t.ADD,
+    payload: {
+      type,
+      props,
+    },
+  };
+};
+
+export default add;
