+++ app/modules/content-items/actions/index.js
@@ -1,0 +1,13 @@
+// @flow
+
+import addToState from './addToState';
+import editInState from './editInState';
+import add from './add';
+import edit from './edit';
+
+export {
+  addToState,
+  editInState,
+  add,
+  edit,
+};
