--- app/modules/content-items/tests/actions.test.js
@@ -1,185 +1,0 @@
-// @flow
-
-import * as t from '../actionTypes';
-import {
-  addToState,
-  editInState,
-  add,
-  edit,
-} from '../actions';
-import { contentItemTypes } from '../model';
-
-describe(`actions`, (): void => {
-
-  describe(`addToState`, (): void => {
-
-    const dummyId = 'abcdefghij';
-    const dummyPlainTextType = contentItemTypes.HEADING;
-    const dummyPlainTextProps = {
-      text: 'Lorem ipsum dolor sit amet.',
-    };
-
-    it(`returns a contentItem ADD_TO_STATE action containing the passed props`, (): void => {
-      const expectedAction: t.AddToStateAction = {
-        type: t.ADD_TO_STATE,
-        payload: {
-          id: dummyId,
-          type: dummyPlainTextType,
-          props: dummyPlainTextProps,
-        },
-      };
-      expect(addToState(dummyId, dummyPlainTextType, dummyPlainTextProps)).toEqual(expectedAction);
-    });
-
-    it(`throws an error, when the passed type is a plainText type and the passed props don't match this type`, (): void => {
-      expect((): any => addToState(
-        dummyId,
-        dummyPlainTextType,
-        {
-          definitelyNotAPlainTextProp: 'abcde',
-        },
-      )).toThrowError(`"text" prop must be defined.`);
-    });
-
-    it(`trims all passed plainText string props, when the passed string props contain unnecessary whitespace`, (): void => {
-      const expectedAction: t.AddToStateAction = {
-        type: t.ADD_TO_STATE,
-        payload: {
-          id: dummyId,
-          type: dummyPlainTextType,
-          props: dummyPlainTextProps,
-        },
-      };
-      expect(addToState(
-        dummyId,
-        dummyPlainTextType,
-        {
-          text: `   ${dummyPlainTextProps.text}   `,
-        },
-      )).toEqual(expectedAction);
-    });
-
-    it(`throws an error, when a non-nullable plainText string prop is an empty string`, (): void => {
-      expect((): any => addToState(
-        dummyId,
-        dummyPlainTextType,
-        {
-          text: '   ',
-        },
-      )).toThrowError(`"text" prop cannot be an empty string.`);
-    });
-
-  });
-
-  describe(`editInState`, (): void => {
-
-    const dummyId = 'abcdefghij';
-    const dummyPlainTextType = contentItemTypes.HEADING;
-    const dummyPlainTextProps = {
-      text: 'Lorem ipsum dolor sit amet.',
-    };
-
-    it(`returns a contentItem EDIT_IN_STATE action containing the passed props`, (): void => {
-      const expectedAction: t.EditInStateAction = {
-        type: t.EDIT_IN_STATE,
-        payload: {
-          id: dummyId,
-          type: dummyPlainTextType,
-          props: dummyPlainTextProps,
-        },
-      };
-      expect(editInState(dummyId, dummyPlainTextType, dummyPlainTextProps)).toEqual(expectedAction);
-    });
-
-    it(`throws an error, when the passed props contain invalid keys for the given contentItemType`, (): void => {
-      expect((): any => editInState(
-        dummyId,
-        dummyPlainTextType,
-        {
-          ...dummyPlainTextProps,
-          definitelyNotAValidProp: 'abcde',
-        },
-      )).toThrowError();
-    });
-
-    it(`throws an error, when all passed props are undefined`, (): void => {
-      expect((): any => editInState(
-        dummyId,
-        dummyPlainTextType,
-        {},
-      )).toThrowError();
-    });
-
-    it(`trims all passed plainText string props, when the passed string props contain unnecessary whitespace`, (): void => {
-      const expectedAction: t.EditInStateAction = {
-        type: t.EDIT_IN_STATE,
-        payload: {
-          id: dummyId,
-          type: dummyPlainTextType,
-          props: dummyPlainTextProps,
-        },
-      };
-      expect(editInState(
-        dummyId,
-        dummyPlainTextType,
-        {
-          text: `   ${dummyPlainTextProps.text}   `,
-        },
-      )).toEqual(expectedAction);
-    });
-
-    it(`throws an error, when a non-nullable plainText string prop is an empty string`, (): void => {
-      expect((): any => editInState(
-        dummyId,
-        dummyPlainTextType,
-        {
-          text: '   ',
-        },
-      )).toThrowError();
-    });
-
-  });
-
-  describe(`add`, (): void => {
-
-    const dummyType = contentItemTypes.HEADING;
-    const dummyProps = {
-      text: 'Lorem ipsum dolor sit amet.',
-    };
-
-    it(`returns a contentItem ADD action containing the passed props`, (): void => {
-      const expectedAction: t.AddAction = {
-        type: t.ADD,
-        payload: {
-          type: dummyType,
-          props: dummyProps,
-        },
-      };
-      expect(add(dummyType, dummyProps)).toEqual(expectedAction);
-    });
-
-  });
-
-  describe(`edit`, (): void => {
-
-    const dummyId = 'abcdefghij';
-    const dummyType = contentItemTypes.HEADING;
-    const dummyProps = {
-      text: 'Lorem ipsum dolor sit amet.',
-    };
-
-    it(`returns a contentItem EDIT action containing the passed props`, (): void => {
-      const expectedAction: t.EditAction = {
-        type: t.EDIT,
-        payload: {
-          id: dummyId,
-          type: dummyType,
-          props: dummyProps,
-        },
-      };
-      expect(edit(dummyId, dummyType, dummyProps)).toEqual(expectedAction);
-    });
-
-  });
-
-});
